requirements.txt
PyQt6
numpy
matplotlib
pyqtgraph
pyhackrf
scipy
